pyproject.toml
[tool.poetry]
name = "bullpen-api"
version = "0.1.0"
description = "The API for Bullpen AI"
authors = ["Kaizen Software <hello@kaizen.software>"]
license = "Proprietary"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.17.0"}
celery = "^5.2.6"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.8"
alembic = "^1.10.2"
psycopg2-binary = "^2.9.6"
requests = "^2.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[tool.poetry.extras]
monitoring = ["flower"]

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.2.0"
coverage = "^7.2.5"
httpx = "^0.24.0"
celery-stubs = "^0.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = "True"
plugins = "sqlalchemy.ext.mypy.plugin"